Use uint64_t vectors for RSA ciphertext
Update rsa-small API to represent ciphertext as std::vector<uint64_t> instead of std::vector<uint8_t>. encrypt and encryptText now return vectors of uint64_t; decrypt and decryptText now accept vectors of uint64_t. This avoids truncation of RSA block values larger than a byte.

+++ src/rsa-small/rsa-small.h
@@ -27,19 +27,19 @@ public:
     static keyPair generate_keys(uint8_t bit_length);
 
     // Binary-oriented API
-    static std::vector<uint8_t> encrypt(const std::vector<uint8_t>& plaintext, const key& publicKey);
-    static std::vector<uint8_t> decrypt(const std::vector<uint8_t>& ciphertext, const key& privateKey);
+    static std::vector<uint64_t> encrypt(const std::vector<uint8_t>& plaintext, const key& publicKey);
+    static std::vector<uint8_t> decrypt(const std::vector<uint64_t>& ciphertext, const key& privateKey);
 
     // Text-oriented API
     /**
      * Encrypts a plaintext string and returns the ciphertext as a vector of bytes.
      */
-    static std::vector<uint8_t> encryptText(const std::string& plaintext, const key& publicKey);
+    static std::vector<uint64_t> encryptText(const std::string& plaintext, const key& publicKey);
 
     /**
      * Decrypts a ciphertext represented as a vector of bytes and returns the resulting plaintext string.
      */
-    static std::string decryptText(const std::vector<uint8_t>& ciphertext, const key& privateKey);
+    static std::string decryptText(const std::vector<uint64_t>& ciphertext, const key& privateKey);
 
 private:
     /**
